added communications folder
added the code which should connect the pi with the local server, it should work but the file of pi client should be in the raspberry pi itself.

with this addition i faced some problem with the labeeb code itself which i tried to fix, some of them were fixed and some are still there, regardless it's not my job but i will try to fix them after i make sure the communications code is working.